ahora se puede calificar a un negocio con las estrellas

+++ database/migrations/2022_11_06_051300_create_calificaciones_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->integer('id_negocio');
             $table->string('nombre_usuario');
             $table->string('comentario');
-            $table->integer('calificacion');
+            $table->float('calificacion');
             $table->timestamps();
         });
     }
